Make version number accessible in code
We're already making versioned releases, so let's make that version
number available in code.

There are many ways to do this, such as release scripts, generating
version numbers from a CMake project version etc. We chose this approach
of manually updating a single header file because it is the simplest we
could think of for a project of so little complexity, and because we
then don't introduce dependencies on new scripting languages etc.

The empty quotes for the `static_assert`s in tests.cpp are there to
avoid a compilation error on g++-5.