Make operator<< enabled only if the skill Printable is present
operator<< is currently always defined for all types. This may have side effects in some cases, like using Boost.Test. It should be generated only when using the Printable skill.

It is now enabled only when the Printable skill is set.

Reordered the skill Printable in underlying_functionalities.hpp so that the operator<< is next to it.

Added a unit test.