Add possibility for trivially constructible types
Previously, NamedType had a user-provided constructor, which prevented
it from being trivially constructible even if the underlying type was.

This will again prevent any class which has a NamedType member from
being trivially constructible.

Let's instead explicitly default the default constructor, which is also
a lot simpler than the previous definition.

To answer some potential questions:

* Why did we remove constexpr? *

Leaving constexpr in there for the explicitly defaulted constructor is
not allowed if the underlying type is not constexpr default
constructible. And leaving it out will automatically do the right thing.
https://timsong-cpp.github.io/cppwp/n4659/dcl.fct.def.default#2:
"An explicitly-defaulted function that is not defined as deleted may be
declared constexpr only if it would have been implicitly declared as
constexpr. If a function is explicitly defaulted on its first
declaration, it is implicitly considered to be constexpr if the implicit
declaration would be."

* Why did we remove the noexcept specifiier? *

Again, for the explicitly defaulted constructor, the default does the
right thing automatically.
https://timsong-cpp.github.io/cppwp/n4659/except.spec#3:
"If a declaration of a function (...) is defaulted on its first
declaration, (...) the exception specfication is as specified below"

"Below" here refers to
https://timsong-cpp.github.io/cppwp/n4659/except.spec#7:
"A constructor without a noexcept-specifier that is defaulted on its
first declaration, has a potentially-throwing exception specification if
and only if any of the following constructs is potentially-throwing:
- a constructor selected by overload resolution in the implicit
  definition of the constructor for class X to initialize a potentially
  constructed subobject, or
- a subexpression of such an initialization, such as a default argument
  expression, or,
- for a default constructor, a default member initializer."

So our constructor will be noexcept if the subobject of the underlying
type is noexcept default constructible. I also added tests for this.