Disable -Weffc++ on gcc 5 and 6
Weffc++ is of questionable usefulness in C++11, especially on older
versions of GCC. See for instance these comments from Jonathan Wakely:

- https://gcc.gnu.org/bugzilla/show_bug.cgi?id=55837
- https://stackoverflow.com/a/11529328/7084

For instance, it looks like it forced an unnecessary manual
initialisation of the member in 191cfeba4d2e2. As of GCC 7, they have
updated -Weffc++ to no longer warn about this.

This change is needed for an upcoming simplification of the default
constructor, which would otherwise start giving this warning (but only
on GCC 5 and 6):

```
fluent::NamedType<int,
____C_A_T_C_H____T_E_S_T____8()::StrongIntTag>::value_’ should be
initialized in the member initialization list [-Werror=effc++]
```

I assume this is the same warning that forced the change in
191cfeba4d2e2.

I suggest we disable -Weffc++ on older versions of GCC, since it seems
it doesn't work very well with modern C++ in those versions. But we
could also stop using -Weffc++ entirely if that's simpler.

I tested this change by adding `VERBOSE=1` to `make` and verifying that
the warning still gets enabled on e.g. G++ 7 and 9 in Travis, but not in
5 and 6.